Few viewer changes (#2553)
* Viewer changes for render panel for fps and resolution

Small changes for the new viewer including changing the FPS slider to a text entry and fixing the incorrect resolution text box description.

* Check if render_time is nonzero in render state machine

Added a check if render_time is non zero in the render_state_machine file in the new viewer due to a divide by zero error occasionally causing the viewer to stop running. I noticed this usually happened when using luxenacto-huge.