Setting save_world_frame as False by default and Splat export panel in viewer (#2700)
* Setting save_world_frame as False by default

Setting save_world_frame as False by default for the exporter as well as in the viewer, since it is exports the point cloud in a different coordinate system from the Blender add-on and there is not a similar checkbox in the original viewer, so by default the exported point cloud would not be compatible. This flag can still be set true in the cmd command and in the new viewer

Also updating the blender docs to mention support for gs scenes

* Adding splat export panel in viewer

Adding the splat option in the export panel which returns a basic splat export cmd. Also renaming the splat exports as splat.ply instead of point_cloud.ply

* Showing export panels based on method

If gaussian splatting is running, then the crop box, mesh, and point cloud exports are hidden/disabled. Else, all but splat export is visible.

* Fix pyright type check

* Changing viewer model argument for export panels

* Update export_panel.py

* Update export_panel.py

* Update export_panel.py

* Update exporter.py

* Update exporter.py

---------